+++ src/Controls/tests/TestCases/Issues/Issue18161.xaml
@@ -1,0 +1,22 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<FlyoutPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+            xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+            x:Class="Maui.Controls.Sample.Issues.Issue18161">
+  <FlyoutPage.Flyout>
+    <ContentPage Title="The FlyOut">
+      <Label Text="This is the flyout" />
+    </ContentPage>
+  </FlyoutPage.Flyout>
+  <FlyoutPage.Detail>
+    <NavigationPage>
+      <x:Arguments>
+        <ContentPage Title="The Detail">
+          <StackLayout>
+            <Label Text="The detail page." />
+            <Button Text="Change Flyout behaviour" AutomationId="ToggleBehaviour" Clicked="ToggleBehaviour_Clicked" />
+          </StackLayout>
+        </ContentPage>
+      </x:Arguments>
+    </NavigationPage>
+  </FlyoutPage.Detail>
+</FlyoutPage>
