+++ src/Controls/tests/TestCases/Issues/Issue20858.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <FlyoutPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-            x:Class="Maui.Controls.Sample.Issues.FlyoutPageTest">
+            x:Class="Maui.Controls.Sample.Issues.Issue20858">
   <FlyoutPage.Flyout>
     <ContentPage Title="The FlyOut">
       <Label Text="This is the flyout" />
@@ -12,9 +12,8 @@
       <x:Arguments>
         <ContentPage Title="The Detail">
           <StackLayout>
-            <Label Text="This is the detail" />
+            <Label Text="Please rotate the device to portrait mode to test. Open the flyout and then rotate device to landscape." />
             <Button Text="Open the Flyout to begin testing" AutomationId="OpenFlyout" Clicked="OpenFlyout_Clicked" />
-            <Button Text="Change Flyout behaviour" AutomationId="ToggleBehaviour" Clicked="ToggleBehaviour_Clicked" />
           </StackLayout>
         </ContentPage>
       </x:Arguments>
